+++ code/src/cljs/sixsq/nuvla/ui/edge/spec.cljs
@@ -23,6 +23,9 @@
 
 (s/def ::nuvlabox-usb-api-key any?)
 
+(s/def ::nuvlabox-ssh-key any?)
+(s/def ::nuvlabox-private-ssh-key string?)
+
 (s/def ::page int?)
 (s/def ::elements-per-page int?)
 (s/def ::total-elements int?)
@@ -34,6 +37,8 @@
 
 (s/def ::vpn-infra any?)
 
+; ssh key association
+(s/def ::ssh-keys-available any?)
 
 (s/def ::db (s/keys :req [::loading?
                           ::nuvlaboxes
@@ -48,20 +53,26 @@
                           ::open-modal
                           ::nuvlabox-created-id
                           ::nuvlabox-usb-api-key
-                          ::vpn-infra]))
+                          ::nuvlabox-ssh-key
+                          ::nuvlabox-private-ssh-key
+                          ::vpn-infra
+                          ::ssh-keys-available]))
 
 
-(def defaults {::loading?            false
-               ::nuvlaboxes          nil
-               ::nuvlabox-releases   nil
-               ::state-nuvlaboxes    nil
-               ::status-nuvlaboxes   nil
-               ::page                1
-               ::elements-per-page   9
-               ::total-elements      0
-               ::full-text-search    nil
-               ::state-selector      nil
-               ::open-modal          nil
-               ::nuvlabox-created-id nil
-               ::nuvlabox-usb-api-key nil
-               ::vpn-infra           nil})
+(def defaults {::loading?                 false
+               ::nuvlaboxes               nil
+               ::nuvlabox-releases        nil
+               ::state-nuvlaboxes         nil
+               ::status-nuvlaboxes        nil
+               ::page                     1
+               ::elements-per-page        9
+               ::total-elements           0
+               ::full-text-search         nil
+               ::state-selector           nil
+               ::open-modal               nil
+               ::nuvlabox-created-id      nil
+               ::nuvlabox-usb-api-key     nil
+               ::nuvlabox-ssh-key         nil
+               ::nuvlabox-private-ssh-key nil
+               ::vpn-infra                nil
+               ::ssh-keys-available       nil})