+++ code/src/cljs/sixsq/nuvla/ui/edge_detail/subs.cljs
@@ -19,6 +19,12 @@
 
 
 (reg-sub
+  ::nuvlabox-ssh-keys
+  (fn [db]
+    (::spec/nuvlabox-ssh-keys db)))
+
+
+(reg-sub
   ::nuvlabox-peripherals
   (fn [db]
     (::spec/nuvlabox-peripherals db)))